conformant: stablize after refactoring.
This is the last big change of the conformant style.
I've tried clean up the style since it's previous code is not much maintained
and it's too much depends on the previous product design.

it removes overlap states (i coudn't find any usage)
make it stable with some test suites.

Change-Id: Iefbe327c3a683a796dd50a6f52afbb5ad8cf35b4